nrh disptacher and detail level atomics
